backend/requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.4.4
wheel==0.42.0